fix: last byte check when deserializing json into bit vector (#36)
After deserializing json into bit vector, we need to ensure the last
byte of the bit vector to have at least `8 - length % 8` leading zeros,
but in current code we are checking the opposite.

Also `@clz(@as(u8, length / 8))` seems wrong. Its value is 8 when length
is between 0 and 7, when the maximum number of leading zeros allowed for
length between 0 and 7 is 7.